backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot==21.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.0
qrcode[pil]==7.4.2
Pillow==10.4.0
python-multipart==0.0.9
